+++ packages/blade/src/components/Typography/Heading/Heading.tsx
@@ -2,17 +2,20 @@
 import type { ReactElement } from 'react';
 import { BaseText } from '../BaseText';
 import type { BaseTextProps } from '../BaseText/types';
+import { useValidateAsProp } from '../utils';
 import type { ColorContrast, ColorContrastTypes, TextTypes } from '~tokens/theme/theme';
 import { getStyledProps } from '~components/Box/styledProps';
 import type { StyledPropsBlade } from '~components/Box/styledProps';
 import type { Theme } from '~components/BladeProvider';
 import type { StringChildrenType, TestID } from '~utils/types';
-import { getPlatformType } from '~utils';
+import { isReactNative } from '~utils';
 
 type HeadingVariant = 'regular' | 'subheading';
 type HeadingSize = 'small' | 'medium' | 'large';
 
+const validAsValues = ['span', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6'] as const;
 type HeadingCommonProps = {
+  as?: typeof validAsValues[number];
   /**
    * Overrides the color of the Heading component.
    *
@@ -60,6 +63,7 @@ export type HeadingProps<T> = T extends {
   : T;
 
 const getProps = <T extends { variant: HeadingVariant }>({
+  as,
   variant,
   size,
   type,
@@ -69,9 +73,8 @@ const getProps = <T extends { variant: HeadingVariant }>({
   testID,
 }: Pick<
   HeadingProps<T>,
-  'variant' | 'size' | 'type' | 'weight' | 'contrast' | 'color' | 'testID'
+  'as' | 'variant' | 'size' | 'type' | 'weight' | 'contrast' | 'color' | 'testID'
 >): Omit<BaseTextProps, 'children'> => {
-  const isPlatformWeb = getPlatformType() === 'browser' || getPlatformType() === 'node';
   const colorContrast: keyof ColorContrast = contrast ? `${contrast!}Contrast` : 'lowContrast';
   const props: Omit<BaseTextProps, 'children'> = {
     color: color ?? `surface.text.${type ?? 'normal'}.${colorContrast}`,
@@ -80,7 +83,7 @@ const getProps = <T extends { variant: HeadingVariant }>({
     fontStyle: 'normal',
     lineHeight: 300,
     fontFamily: 'text',
-    accessibilityProps: isPlatformWeb ? {} : { role: 'heading' },
+    accessibilityProps: isReactNative() ? { role: 'heading' } : {},
     componentName: 'heading',
     testID,
   };
@@ -89,15 +92,15 @@ const getProps = <T extends { variant: HeadingVariant }>({
     if (!size || size === 'small') {
       props.fontSize = 200;
       props.lineHeight = 300;
-      props.as = isPlatformWeb ? 'h6' : undefined;
+      props.as = 'h6';
     } else if (size === 'medium') {
       props.fontSize = 300;
       props.lineHeight = 200;
-      props.as = isPlatformWeb ? 'h5' : undefined;
+      props.as = 'h5';
     } else if (size === 'large') {
       props.fontSize = 400;
       props.lineHeight = 400;
-      props.as = isPlatformWeb ? 'h4' : undefined;
+      props.as = 'h4';
     }
   } else if (variant === 'subheading') {
     if (weight === 'regular') {
@@ -110,13 +113,16 @@ const getProps = <T extends { variant: HeadingVariant }>({
     }
     props.fontSize = 75;
     props.lineHeight = 50;
-    props.as = isPlatformWeb ? 'h6' : undefined;
+    props.as = 'p';
   }
 
+  // override the computed `as` prop if user passed an `as` prop
+  props.as = as || props.as;
   return props;
 };
 
 export const Heading = <T extends { variant: HeadingVariant }>({
+  as,
   variant = 'regular',
   size, // Not setting default value since the `size` should be undefined with variant="subheading"
   type = 'normal',
@@ -128,7 +134,10 @@ export const Heading = <T extends { variant: HeadingVariant }>({
   textAlign,
   ...styledProps
 }: HeadingProps<T>): ReactElement => {
-  const props = getProps({ variant, size, type, weight, color, contrast, testID });
+  useValidateAsProp({ componentName: 'Heading', as, validAsValues });
+
+  const props = getProps({ as, variant, size, type, weight, color, contrast, testID });
+
   return (
     <BaseText {...props} textAlign={textAlign} {...getStyledProps(styledProps)}>
       {children}